Refactored FeatureFlag annotation to use Feature enum
